+++ lib/domain/models/settings.dart
@@ -23,7 +23,6 @@ class Settings with _$Settings {
     required String languageSeed,
     required bool firstLaunch,
     required bool showBalances,
-    required bool testnetEnabled,
     required bool activeRPCServer,
     required int mainScreenCurrentPage,
     required bool showPriceChart,
@@ -40,7 +39,6 @@ class Settings with _$Settings {
         primaryCurrency:
             AvailablePrimaryCurrency(AvailablePrimaryCurrencyEnum.native),
         showBalances: true,
-        testnetEnabled: false,
         showPriceChart: true,
         priceChartIntervalOption: aedappfm.MarketPriceHistoryInterval.hour,
       );
