+++ lib/ui/menu/settings/customization_menu_view.dart
@@ -43,7 +43,6 @@ class CustomizationMenuView extends ConsumerWidget
     final localizations = AppLocalizations.of(context)!;
 
     final primaryCurrency = ref.watch(selectedPrimaryCurrencyProvider);
-    final environment = ref.watch(environmentProvider);
     return DecoratedBox(
       decoration: BoxDecoration(
         gradient: LinearGradient(
@@ -77,10 +76,6 @@ class CustomizationMenuView extends ConsumerWidget
                       const _SettingsListItem.spacer(),
                       const _ShowPriceChartSettingsListItem(),
                       const _SettingsListItem.spacer(),
-                      if (environment == aedappfm.Environment.mainnet)
-                        const _TestnetEnabledSettingsListItem(),
-                      if (environment == aedappfm.Environment.mainnet)
-                        const _SettingsListItem.spacer(),
                     ],
                   ),
                 ],
@@ -149,27 +144,6 @@ class _ShowPriceChartSettingsListItem extends ConsumerWidget {
       isSwitched: showPriceChart,
       onChanged: (showPriceChart) async {
         await preferencesNotifier.setShowPriceChart(showPriceChart);
-      },
-    );
-  }
-}
-
-class _TestnetEnabledSettingsListItem extends ConsumerWidget {
-  const _TestnetEnabledSettingsListItem();
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    final localizations = AppLocalizations.of(context)!;
-    final testnetEnabledSetting = ref.watch(
-      SettingsProviders.settings.select((settings) => settings.testnetEnabled),
-    );
-    final preferencesNotifier = ref.read(SettingsProviders.settings.notifier);
-    return _SettingsListItem.withSwitch(
-      heading: localizations.testnetEnabled,
-      icon: Symbols.deployed_code_account,
-      isSwitched: testnetEnabledSetting,
-      onChanged: (testnetEnabled) async {
-        await preferencesNotifier.setTestnetEnabled(testnetEnabled);
       },
     );
   }
