+++ lib/ui/menu/settings/settings_sheet.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 import 'package:aewallet/application/account/accounts_notifier.dart';
 import 'package:aewallet/application/authentication/authentication.dart';
 import 'package:aewallet/application/connectivity_status.dart';
@@ -6,7 +8,9 @@ import 'package:aewallet/application/settings/language.dart';
 import 'package:aewallet/application/settings/primary_currency.dart';
 import 'package:aewallet/application/settings/settings.dart';
 import 'package:aewallet/application/settings/version.dart';
+import 'package:aewallet/application/usecases.dart';
 import 'package:aewallet/domain/repositories/features_flags.dart';
+import 'package:aewallet/domain/usecases/new_keychain.usecase.dart';
 import 'package:aewallet/infrastructure/datasources/tokens_list.hive.dart';
 import 'package:aewallet/infrastructure/rpc/websocket_server.dart';
 import 'package:aewallet/model/authentication_method.dart';
@@ -15,10 +19,12 @@ import 'package:aewallet/model/device_lock_timeout.dart';
 import 'package:aewallet/model/privacy_mask_option.dart';
 import 'package:aewallet/model/setting_item.dart';
 import 'package:aewallet/modules/aeswap/application/session/provider.dart';
-import 'package:aewallet/ui/menu/settings/environment_change.dart';
+import 'package:aewallet/modules/aeswap/ui/views/util/app_styles.dart';
 import 'package:aewallet/ui/themes/archethic_theme.dart';
 import 'package:aewallet/ui/themes/archethic_theme_base.dart';
 import 'package:aewallet/ui/themes/styles.dart';
+import 'package:aewallet/ui/util/accounts_dialog.dart';
+import 'package:aewallet/ui/util/ui_util.dart';
 import 'package:aewallet/ui/views/authenticate/auth_factory.dart';
 import 'package:aewallet/ui/views/dapps_board/layouts/dapps_board_sheet.dart';
 import 'package:aewallet/ui/views/main/components/sheet_appbar.dart';
@@ -30,6 +36,7 @@ import 'package:aewallet/ui/widgets/components/icon_widget.dart';
 import 'package:aewallet/ui/widgets/components/sheet_skeleton.dart';
 import 'package:aewallet/ui/widgets/components/sheet_skeleton_interface.dart';
 import 'package:aewallet/ui/widgets/dialogs/authentification_method_dialog.dart';
+import 'package:aewallet/ui/widgets/dialogs/environment_dialog.dart';
 import 'package:aewallet/ui/widgets/dialogs/language_dialog.dart';
 import 'package:aewallet/ui/widgets/dialogs/lock_timeout_dialog.dart';
 import 'package:aewallet/ui/widgets/dialogs/primary_currency_dialog.dart';
@@ -40,6 +47,7 @@ import 'package:aewallet/util/mnemonics.dart';
 import 'package:aewallet/util/universal_platform.dart';
 import 'package:archethic_dapp_framework_flutter/archethic_dapp_framework_flutter.dart'
     as aedappfm;
+import 'package:archethic_lib_dart/archethic_lib_dart.dart' as archethic;
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
