+++ lib/ui/views/intro/layouts/intro_new_wallet_get_first_infos.dart
@@ -138,11 +138,6 @@ class _IntroNewWalletDisclaimerState
                       await ref
                           .read(SettingsProviders.settings.notifier)
                           .setEnvironment(environment as aedappfm.Environment);
-                      await ref
-                          .read(SettingsProviders.settings.notifier)
-                          .setTestnetEnabled(
-                            environment != aedappfm.Environment.mainnet,
-                          );
                     }
 
                     FocusScope.of(context).requestFocus(nameFocusNode);
